Split SwftUI elements from Models
