+++ BusTracker/Models/Stop/BusStopType.swift
@@ -6,34 +6,9 @@
 //
 
 import Foundation
-import SwiftUI
 
-enum BusStopType: String, Decodable {
+enum BusStopType: String, Decodable, CaseIterable {
     case marked = "MKD"
     case custom = "CUS"
     case hailAndRide = "HAR"
-    
-    private func imageView(image: String, height: CGFloat) -> some View {
-        return ZStack {
-            Circle().fill(.busBlue).frame(width: height, height: height)
-            Image(systemName: "circle").resizable().frame(width: height, height: height)
-                .foregroundStyle(.offWhite)
-            Image(systemName: image).resizable().scaledToFit().frame(width: height * 0.7, height: height * 0.7)
-                .foregroundStyle(.offWhite)
-        }
-    }
-    
-    @ViewBuilder
-    func view(height: CGFloat = 16) -> some View {
-        switch self {
-        case .marked:
-            imageView(image: "signpost.and.arrowtriangle.up.fill", height: height)
-        case .hailAndRide:
-            imageView(image: "figure.wave", height: height)
-        case .custom:
-            imageView(image: "mappin", height: height)
-        }
-    }
-    
 }
-
