+++ BusTracker/Views/Operators/OperatorColour.swift
